refactor editor.tsx
