+++ src/components/Editor.tsx
@@ -1,38 +1,18 @@
 import "./index.css";
-import React, { useEffect, useState, useCallback, useRef } from "react";
+import React, { useEffect, useState } from "react";
 import { LoadingScreen } from "../internal/puck/components/LoadingScreen.tsx";
 import { Toaster } from "../internal/puck/ui/Toaster.tsx";
-import { TemplateMetadata } from "../internal/types/templateMetadata.ts";
-import { type Config, InitialHistory } from "@measured/puck";
-import {
-  useMessageSenders,
-  useReceiveMessage,
-} from "../internal/hooks/useMessage.ts";
-import { LayoutSaveState } from "../internal/types/saveState.ts";
+import { type Config } from "@measured/puck";
+import { useMessageSenders } from "../internal/hooks/useMessage.ts";
 import "@measured/puck/puck.css";
 import { DevLogger } from "../utils/devLogger.ts";
 import { ThemeConfig } from "../utils/themeResolver.ts";
-import { ThemeSaveState } from "../internal/types/themeSaveState.ts";
-import { updateThemeInEditor } from "../utils/applyTheme.ts";
-import { InternalThemeEditor } from "../internal/components/InternalThemeEditor.tsx";
-import { InternalLayoutEditor } from "../internal/components/InternalLayoutEditor.tsx";
 import { useQuickFindShortcut } from "../internal/hooks/useQuickFindShortcut.ts";
-import { useLocalStorage } from "../internal/hooks/useLocalStorage.ts";
+import { useMessageReceivers } from "../hooks/useMessageReceivers.ts";
+import { LayoutEditor } from "../internal/components/LayoutEditor.tsx";
+import { ThemeEditor } from "../internal/components/ThemeEditor.tsx";
 
-export const Role = {
-  GLOBAL: "global",
-  INDIVIDUAL: "individual",
-};
-
-export const TARGET_ORIGINS = [
-  "http://localhost",
-  "https://dev.yext.com",
-  "https://qa.yext.com",
-  "https://sandbox.yext.com",
-  "https://www.yext.com",
-  "https://app-qa.eu.yext.com",
-  "https://app.eu.yext.com",
-];
+const devLogger = new DevLogger();
 
 export type EditorProps = {
   document: any;
@@ -40,63 +20,34 @@ export type EditorProps = {
   themeConfig?: ThemeConfig;
 };
 
-const devLogger = new DevLogger();
-
 export const Editor = ({
   document,
   componentRegistry,
   themeConfig,
 }: EditorProps) => {
-  const [puckInitialHistory, setPuckInitialHistory] = useState<
-    InitialHistory | undefined
-  >();
-  const [puckInitialHistoryFetched, setPuckInitialHistoryFetched] =
-    useState<boolean>(false);
-  const [visualConfigurationData, setVisualConfigurationData] = useState<any>(); // json data
-  const [visualConfigurationDataFetched, setVisualConfigurationDataFetched] =
-    useState<boolean>(false); // needed because visualConfigurationData can be empty
-  const [saveState, setSaveState] = useState<LayoutSaveState>();
-  const [saveStateFetched, setSaveStateFetched] = useState<boolean>(false); // needed because saveState can be empty
-  const [devPageSets, setDevPageSets] = useState<any>(undefined);
-  const [devSiteStream, setDevSiteStream] = useState<any>(undefined);
-  const [templateMetadata, setTemplateMetadata] = useState<TemplateMetadata>();
-  const [puckConfig, setPuckConfig] = useState<Config>();
-  const [parentLoaded, setParentLoaded] = useState<boolean>(false);
-
-  // theming
-  const [themeInitialHistory, setThemeInitialHistory] =
-    useState<ThemeSaveState>({ history: [{}], index: 0 });
-  const [themeInitialHistoryFetched, setThemeInitialHistoryFetched] =
-    useState<boolean>(false);
-  const [themeData, setThemeData] = useState<any>(); // json data
-  const [themeDataFetched, setThemeDataFetched] = useState<boolean>(false); // needed because themeData can be empty
-  const [themeSaveState, setThemeSaveState] = useState<
-    ThemeSaveState | undefined
-  >(undefined);
-  const [themeSaveStateFetched, setThemeSaveStateFetched] =
-    useState<boolean>(false); // needed because themeSaveState can be empty
-
   if (document) {
     devLogger.logData("DOCUMENT", document);
   }
 
+  const [devPageSets, setDevPageSets] = useState<any>(undefined);
+  const [devSiteStream, setDevSiteStream] = useState<any>(undefined);
+  const [parentLoaded, setParentLoaded] = useState<boolean>(false);
+
   const {
-    iFrameLoaded,
-    sendDevLayoutSaveStateData,
-    pushPageSets,
-    saveLayoutSaveState,
-    publishVisualConfiguration,
-    saveThemeSaveState,
-    publishThemeConfiguration,
-    sendDevThemeSaveStateData,
-    deleteLayoutSaveState,
-    deleteThemeSaveState,
-  } = useMessageSenders();
-  const {
-    buildVisualConfigLocalStorageKey,
-    buildThemeLocalStorageKey,
-    clearLocalStorage,
-  } = useLocalStorage(templateMetadata);
+    templateMetadata,
+    puckConfig,
+    visualConfigurationData,
+    visualConfigurationDataFetched,
+    layoutSaveState,
+    layoutSaveStateFetched,
+    themeData,
+    themeDataFetched,
+    themeSaveState,
+    themeSaveStateFetched,
+  } = useMessageReceivers(componentRegistry);
+
+  const { pushPageSets } = useMessageSenders();
+
   useQuickFindShortcut();
 
   // redirect to 404 page when going to /edit page outside of Storm
@@ -117,17 +68,6 @@ export const Editor = ({
     }
   }, []);
 
-  useReceiveMessage("getTemplateMetadata", TARGET_ORIGINS, (send, payload) => {
-    const puckConfig = componentRegistry.get(payload.templateId);
-    setPuckConfig(puckConfig);
-    const templateMetadata = payload as TemplateMetadata;
-    setTemplateMetadata(payload as TemplateMetadata);
-    devLogger.enable(templateMetadata.isxYextDebug);
-    devLogger.logData("TEMPLATE_METADATA", templateMetadata);
-    devLogger.logData("PUCK_CONFIG", puckConfig);
-    send({ status: "success", payload: { message: "payload received" } });
-  });
-
   useEffect(() => {
     if (templateMetadata?.isDevMode) {
       try {
@@ -143,59 +83,6 @@ export const Editor = ({
   }, [templateMetadata?.isDevMode]);
 
   useEffect(() => {
-    iFrameLoaded({ payload: { message: "iFrame is loaded" } });
-  }, []);
-
-  useReceiveMessage("getSaveState", TARGET_ORIGINS, (send, payload) => {
-    devLogger.logData("SAVE_STATE", payload);
-    setSaveState(payload as LayoutSaveState);
-    setSaveStateFetched(true);
-    send({ status: "success", payload: { message: "saveState received" } });
-  });
-
-  useReceiveMessage("getThemeSaveState", TARGET_ORIGINS, (send, payload) => {
-    const themeSaveState = {
-      history: payload?.history
-        ? jsonFromEscapedJsonString(payload?.history)
-        : [],
-      index: payload?.index ?? 0,
-    };
-    devLogger.logData("THEME_SAVE_STATE", payload);
-    setThemeSaveState(themeSaveState);
-    setThemeSaveStateFetched(true);
-    send({
-      status: "success",
-      payload: { message: "themeSaveState received" },
-    });
-  });
-
-  useReceiveMessage(
-    "getVisualConfigurationData",
-    TARGET_ORIGINS,
-    (send, payload) => {
-      const vcd = jsonFromEscapedJsonString(payload.visualConfigurationData);
-      devLogger.logData("VISUAL_CONFIGURATION_DATA", vcd);
-      setVisualConfigurationData(vcd);
-      setVisualConfigurationDataFetched(true);
-      send({
-        status: "success",
-        payload: { message: "getVisualConfigurationData received" },
-      });
-    }
-  );
-
-  useReceiveMessage("getThemeData", TARGET_ORIGINS, (send, payload) => {
-    const themeData = jsonFromEscapedJsonString(payload as unknown as string);
-    devLogger.logData("THEME_DATA", themeData);
-    setThemeData(themeData);
-    setThemeDataFetched(true);
-    send({
-      status: "success",
-      payload: { message: "getThemeData received" },
-    });
-  });
-
-  useEffect(() => {
     if (
       typeof window !== "undefined" &&
       templateMetadata?.isDevMode &&
@@ -207,387 +94,44 @@ export const Editor = ({
     }
   }, [templateMetadata?.isDevMode, devPageSets]);
 
-  /**
-   * Clears localStorage and resets the save data in the DB
-   */
-  const clearHistory = () => {
-    devLogger.logFunc("clearHistory");
-    clearLocalStorage();
-    if (templateMetadata?.isThemeMode) {
-      deleteThemeSaveState();
-    } else {
-      deleteLayoutSaveState();
-    }
-  };
-
-  const isFirstRender = useRef(true);
-  useEffect(() => {
-    if (isFirstRender.current) {
-      isFirstRender.current = false; // toggle flag after first render/mounting
-      return;
-    }
-    loadPuckInitialHistory();
-    loadThemeInitialHistory();
-  }, [
-    templateMetadata,
-    saveStateFetched,
-    visualConfigurationDataFetched,
-    themeDataFetched,
-    themeSaveStateFetched,
-  ]);
-
-  /**
-   * Determines the initialHistory to send to Puck. It is based on a combination
-   * of localStorage and saveState (from the DB).
-   *
-   * When in dev mode, only use localStorage if it exists.
-   * When NOT dev mode:
-   * 1. if no saveState, clear localStorage and use nothing
-   * 2. if saveState, find matching hash in localStorage:
-   *  - if match, use localStorage with index set to saveState hash
-   *  - if no match, use saveState directly and clear localStorage
-   */
-  const loadPuckInitialHistory = useCallback(() => {
-    if (
-      !visualConfigurationDataFetched ||
-      !saveStateFetched ||
-      !templateMetadata
-    ) {
-      return;
-    }
-
-    devLogger.logFunc("loadPuckInitialHistory");
-
-    // Only load Content data for theme mode
-    if (templateMetadata.isThemeMode) {
-      if (visualConfigurationData) {
-        setPuckInitialHistory({
-          histories: [{ id: "root", state: { data: visualConfigurationData } }],
-          appendData: false,
-        });
-      }
-      setPuckInitialHistoryFetched(true);
-      return;
-    }
-
-    if (templateMetadata.isDevMode && !templateMetadata.devOverride) {
-      // Check localStorage for existing Puck history
-      const localHistoryArray = window.localStorage.getItem(
-        buildVisualConfigLocalStorageKey()
-      );
-
-      // Use localStorage directly if it exists
-      if (localHistoryArray) {
-        devLogger.log("Layout Dev Mode - Using layout data from localStorage");
-        const localHistories = JSON.parse(localHistoryArray);
-        const localHistoryIndex = localHistories.length - 1;
-        setPuckInitialHistory({
-          histories: localHistories,
-          index: localHistoryIndex,
-          appendData: false,
-        });
-        setPuckInitialHistoryFetched(true);
-        return;
-      }
-
-      // Otherwise start fresh from Content
-      devLogger.log(
-        "Layout Dev Mode - No localStorage. Using layout data from Content"
-      );
-      if (visualConfigurationData) {
-        setPuckInitialHistory({
-          histories: [{ id: "root", state: { data: visualConfigurationData } }],
-          appendData: false,
-        });
-      }
-      setPuckInitialHistoryFetched(true);
-
-      return;
-    }
-
-    // Nothing in save_state table, start fresh from Content
-    if (!saveState) {
-      clearLocalStorage();
-
-      devLogger.log(
-        "Layout Prod Mode - No saveState. Using layout data from Content"
-      );
-      if (visualConfigurationData) {
-        setPuckInitialHistory({
-          histories: [{ id: "root", state: { data: visualConfigurationData } }],
-          appendData: false,
-        });
-      }
-      setPuckInitialHistoryFetched(true);
-
-      return;
-    }
-
-    // Check localStorage for existing Puck history
-    const localHistoryArray = window.localStorage.getItem(
-      buildVisualConfigLocalStorageKey()
-    );
-
-    // No localStorage, start from saveState
-    if (!localHistoryArray) {
-      devLogger.log(
-        "Layout Prod Mode - No localStorage. Using layout saveState"
-      );
-      setPuckInitialHistory({
-        histories: visualConfigurationData
-          ? [
-              { id: "root", state: { data: visualConfigurationData } },
-              {
-                id: saveState.hash,
-                state: jsonFromEscapedJsonString(saveState.history),
-              },
-            ]
-          : [
-              {
-                id: saveState.hash,
-                state: jsonFromEscapedJsonString(saveState.history),
-              },
-            ],
-        index: visualConfigurationData ? 1 : 0,
-        appendData: false,
-      });
-      setPuckInitialHistoryFetched(true);
-      return;
-    }
-
-    const localHistoryIndex = JSON.parse(localHistoryArray).findIndex(
-      (item: any) => item.id === saveState?.hash
-    );
-
-    // If local storage reset Puck history to it
-    if (localHistoryIndex !== -1) {
-      devLogger.log(
-        "Layout Prod Mode - Using localStorage visual configuration"
-      );
-      setPuckInitialHistory({
-        histories: JSON.parse(localHistoryArray),
-        index: localHistoryIndex,
-        appendData: false,
-      });
-      setPuckInitialHistoryFetched(true);
-      return;
-    }
-
-    // otherwise start fresh - this user doesn't have localStorage that reflects the saved state
-    clearLocalStorage();
-  }, [
-    setPuckInitialHistory,
-    setPuckInitialHistoryFetched,
-    clearLocalStorage,
-    buildVisualConfigLocalStorageKey,
-  ]);
-
-  // Handles sending the devSaveStateData on initial load so that nothing isn't rendered for preview.
-  // Subsequent changes are sent via handleHistoryChange in InternalLayoutEditor.tsx and handleThemeChange
-  // in InternalThemeEditor.tsx
-  useEffect(() => {
-    if (!puckInitialHistoryFetched || templateMetadata?.isThemeMode) {
-      return;
-    }
-
-    const historyToSend =
-      puckInitialHistory?.histories[puckInitialHistory.histories.length - 1]
-        .state.data;
-
-    devLogger.logFunc("sendDevSaveStateData useEffect");
-    sendDevLayoutSaveStateData({
-      payload: { devSaveStateData: JSON.stringify(historyToSend) },
-    });
-  }, [puckInitialHistoryFetched, puckInitialHistory, templateMetadata]);
-
-  /*
-   * Determines the initial theme data to send to the editor.
-   * Prod mode: If DB theme_save_state exists, use that. Otherwise, pull from Content
-   * Dev mode: If localstorage theme data exists, use that. Otherwise, pull from Content
-   */
-  const loadThemeInitialHistory = useCallback(() => {
-    if (
-      !themeSaveStateFetched ||
-      !templateMetadata ||
-      !templateMetadata.isThemeMode
-    ) {
-      return;
-    }
-    devLogger.logFunc("loadThemeHistory");
-    devLogger.logData("THEME_DATA", themeData);
-    devLogger.logData("THEME_SAVE_STATE", themeSaveState);
-
-    if (templateMetadata.isDevMode && !templateMetadata.devOverride) {
-      // Check localStorage for existing theme history
-      const localHistoryArray = window.localStorage.getItem(
-        buildThemeLocalStorageKey()
-      );
-
-      // Use localStorage directly if it exists
-      if (localHistoryArray) {
-        devLogger.log("Theme Dev Mode - Using theme localStorage");
-        const localHistories = JSON.parse(localHistoryArray);
-        const localHistoryIndex = localHistories.length - 1;
-        setThemeInitialHistory({
-          history: localHistories,
-          index: localHistoryIndex,
-        });
-        setThemeInitialHistoryFetched(true);
-        return;
-      }
-
-      // Otherwise start fresh from Content
-      devLogger.log(
-        "Theme Dev Mode - No localStorage. Using theme data from Content"
-      );
-      if (themeData) {
-        setThemeInitialHistory({
-          history: [themeData],
-          index: 0,
-        });
-      }
-      setThemeInitialHistoryFetched(true);
-      return;
-    }
-
-    // Nothing in theme_save_state table, start fresh from Content
-    if (!themeSaveState) {
-      clearLocalStorage();
-
-      devLogger.log(
-        "Theme Prod Mode - No saveState. Using theme data from Content"
-      );
-      if (themeData) {
-        setThemeInitialHistory({
-          history: [themeData],
-          index: 0,
-        });
-      }
-      setThemeInitialHistoryFetched(true);
-
-      return;
-    }
-
-    // themeSaveState exists, combine with themeData from content
-    if (themeData) {
-      devLogger.log("Theme Prod Mode - Using themeSaveState and themeData");
-      const history = themeData
-        ? [themeData, ...themeSaveState.history]
-        : [...themeSaveState.history];
-      const themeInitialHistory = {
-        history: history,
-        index: history.length - 1,
-      };
-      setThemeInitialHistory(themeInitialHistory);
-      setThemeInitialHistoryFetched(true);
-      return;
-    }
-
-    devLogger.log(
-      "No loadThemeInitialHistory case matched, setting to empty theme."
-    );
-    setThemeInitialHistory({
-      history: [{}],
-      index: 0,
-    });
-    setThemeInitialHistoryFetched(true);
-  }, [
-    setThemeInitialHistory,
-    setThemeInitialHistoryFetched,
-    clearLocalStorage,
-    buildThemeLocalStorageKey,
-  ]);
-
-  // Log PUCK_INITIAL_HISTORY (layout) on load
-  useEffect(() => {
-    if (puckInitialHistory) {
-      devLogger.logData("PUCK_INITIAL_HISTORY", puckInitialHistory);
-    }
-  }, [puckInitialHistory]);
-
-  // Log THEME_INITIAL_HISTORY on load and update theme in editor to reflect save state
-  useEffect(() => {
-    devLogger.logData("THEME_INITIAL_HISTORY", themeInitialHistory);
-    if (themeInitialHistory && themeConfig && templateMetadata?.isThemeMode) {
-      updateThemeInEditor(
-        themeInitialHistory.history[themeInitialHistory.index],
-        themeConfig
-      );
-    }
-  }, [themeInitialHistory, themeConfig, templateMetadata?.isThemeMode]);
-
-  // Handles sending the sendDevThemeSaveStateData on initial load so that nothing isn't rendered for preview.
-  // Subsequent changes are sent via handleChange in ThemeSidebar.tsx.
-  useEffect(() => {
-    if (!themeInitialHistoryFetched || !templateMetadata?.isThemeMode) {
-      return;
-    }
-
-    const themeToSend = themeInitialHistory?.history[themeInitialHistory.index];
-
-    devLogger.logFunc("sendDevThemeSaveStateData useEffect");
-    sendDevThemeSaveStateData({
-      payload: { devThemeSaveStateData: JSON.stringify(themeToSend) },
-    });
-  }, [themeInitialHistoryFetched, themeInitialHistory, templateMetadata]);
-
   const isLoading =
     !puckConfig ||
     !templateMetadata ||
     !document ||
-    !saveStateFetched ||
+    !(layoutSaveStateFetched || templateMetadata?.isThemeMode) ||
     !visualConfigurationDataFetched ||
-    !puckInitialHistoryFetched ||
-    (templateMetadata.isThemeMode &&
-      (!themeDataFetched ||
-        !themeSaveStateFetched ||
-        !themeInitialHistoryFetched));
+    !(themeDataFetched || !templateMetadata?.isThemeMode) ||
+    !(themeSaveStateFetched || !templateMetadata?.isThemeMode);
 
   const progress: number =
-    (100 * // @ts-expect-error adding bools is fine
-      (!!puckConfig +
-        !!templateMetadata +
-        !!document +
-        saveStateFetched +
-        visualConfigurationDataFetched +
-        puckInitialHistoryFetched +
-        (templateMetadata?.isThemeMode // @ts-expect-error adding bools is fine
-          ? themeDataFetched +
-            themeSaveStateFetched +
-            themeInitialHistoryFetched
-          : 0))) /
-    (templateMetadata?.isThemeMode ? 9 : 6);
+    90 * // @ts-expect-error adding bools is fine
+    ((!!puckConfig +
+      !!templateMetadata +
+      !!document +
+      (layoutSaveStateFetched || templateMetadata?.isThemeMode) +
+      visualConfigurationDataFetched +
+      (themeDataFetched || !templateMetadata?.isThemeMode) +
+      (themeSaveStateFetched || !templateMetadata?.isThemeMode)) /
+      7);
 
   return (
     <>
       {!isLoading ? (
         templateMetadata.isThemeMode ? (
-          <InternalThemeEditor
+          <ThemeEditor
             puckConfig={puckConfig}
-            puckInitialHistory={puckInitialHistory}
-            isLoading={isLoading}
             templateMetadata={templateMetadata}
-            publishThemeConfiguration={publishThemeConfiguration}
+            themeSaveState={themeSaveState}
+            themeData={themeData}
+            visualConfigurationData={visualConfigurationData}
             themeConfig={themeConfig}
-            saveThemeSaveState={saveThemeSaveState}
-            themeHistory={themeInitialHistory}
-            setThemeHistory={setThemeInitialHistory}
-            clearThemeHistory={clearHistory}
-            sendDevThemeSaveStateData={sendDevThemeSaveStateData}
-            buildThemeLocalStorageKey={buildThemeLocalStorageKey}
           />
         ) : (
-          <InternalLayoutEditor
+          <LayoutEditor
             puckConfig={puckConfig}
-            puckInitialHistory={puckInitialHistory}
-            isLoading={isLoading}
-            clearHistory={clearHistory}
             templateMetadata={templateMetadata}
-            layoutSaveState={saveState}
-            saveLayoutSaveState={saveLayoutSaveState}
-            publishVisualConfiguration={publishVisualConfiguration}
-            sendDevSaveStateData={sendDevLayoutSaveStateData}
-            buildVisualConfigLocalStorageKey={buildVisualConfigLocalStorageKey}
+            layoutSaveState={layoutSaveState}
+            visualConfigurationData={visualConfigurationData}
           />
         )
       ) : (
@@ -596,8 +140,4 @@ export const Editor = ({
       <Toaster closeButton richColors />
     </>
   );
-};
-
-const jsonFromEscapedJsonString = (escapedJsonString: string) => {
-  return JSON.parse(escapedJsonString.replace(/\\"/g, '"'));
 };
